Greatly simplified the main traffic fragment by extracting a view model.
