+++ build.gradle
@@ -10,12 +10,13 @@ buildscript {
         junitVersion = '5.4.2'
         kotlinVersion = '1.3.50'
         leakcanary = '2.0-beta-2'
-        material = '1.1.0-alpha09'
+        material = '1.1.0-alpha10'
         mockkVersion = '1.9.3'
         okhttp3Version = '3.12.2'
         retrofitVersion = '2.5.0'
         roomVersion = '2.1.0'
-        supportLibVersion = '1.1.0-rc01'
+        supportLibVersion = '1.1.0'
+        viewModelVersion = '2.1.0'
     }
 
     repositories {
@@ -25,7 +26,7 @@ buildscript {
     }
 
     dependencies {
-        classpath "com.android.tools.build:gradle:$androidGradleVersion"
+        classpath "com.android.tools.build:gradle:${androidGradleVersion}"
         classpath "com.github.dcendents:android-maven-gradle-plugin:$androidMavenGradleVersion"
         classpath "com.jfrog.bintray.gradle:gradle-bintray-plugin:$gradleBintrayVersion"
         classpath "de.mannodermaus.gradle.plugins:android-junit5:$junitGradlePluignVersion"
