+++ library/build.gradle
@@ -63,6 +63,7 @@ dependencies {
     testImplementation "org.junit.jupiter:junit-jupiter-params:$junitVersion"
     testImplementation "io.mockk:mockk:$mockkVersion"
 
+    implementation "androidx.lifecycle:lifecycle-extensions:$viewModelVersion"
     implementation "androidx.room:room-runtime:$roomVersion"
     kapt "androidx.room:room-compiler:$roomVersion"
 }
