+++ library/src/main/java/com/chuckerteam/chucker/internal/ui/transaction/TrafficFragment.kt
@@ -2,30 +2,33 @@ package com.chuckerteam.chucker.internal.ui.transaction
 
 import android.content.Context
 import android.os.Bundle
-import android.text.TextUtils
 import android.text.method.LinkMovementMethod
 import android.view.*
 import android.widget.TextView
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.widget.SearchView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.DividerItemDecoration.VERTICAL
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.chuckerteam.chucker.R
-import com.chuckerteam.chucker.internal.data.entity.HttpTransactionTuple
 import com.chuckerteam.chucker.internal.data.repository.RepositoryProvider
 import com.chuckerteam.chucker.internal.support.NotificationHelper
 
 class TrafficFragment : Fragment() {
-    private lateinit var dataSource: LiveData<List<HttpTransactionTuple>>
+    private lateinit var trafficVM: TrafficViewModel
     private lateinit var tutorialView : View
-    private var currentFilter: String? = ""
     private val trafficAdapter = TrafficAdapter { id, _, _ ->
         TransactionActivity.start(activity, id)
+    }
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        trafficVM = ViewModelProviders.of(this)[TrafficViewModel::class.java]
+        trafficVM.executeCurrentQuery()
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -45,11 +48,10 @@ class TrafficFragment : Fragment() {
         }
         tutorialView = findViewById(R.id.tutorial)
         findViewById<TextView>(R.id.link).movementMethod = LinkMovementMethod.getInstance()
-    }
-
-    override fun onAttach(context: Context) {
-        super.onAttach(context)
-        dataSource = getDataSource(currentFilter).andListen()
+        trafficVM.networkTraffic.observe(this@TrafficFragment, Observer { list ->
+            tutorialView.visibility = if (list.isNullOrEmpty()) View.VISIBLE else View.GONE
+            trafficAdapter.submitList(list)
+        })
     }
 
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
@@ -60,7 +62,8 @@ class TrafficFragment : Fragment() {
         with(searchMenuItem.actionView as SearchView) {
             setIconifiedByDefault(true)
             setOnQueryTextListener(object : SearchView.OnQueryTextListener {
-                override fun onQueryTextChange(newText: String?) = executeQuery(newText)
+                override fun onQueryTextChange(newText: String?) =
+                    trafficVM.executeQuery(newText)
                 override fun onQueryTextSubmit(query: String?) = true
             })
         }
@@ -85,35 +88,4 @@ class TrafficFragment : Fragment() {
         .setNegativeButton(R.string.chucker_cancel, null)
         .show()
 
-    private fun executeQuery(newText: String?): Boolean {
-        currentFilter = newText
-        dataSource.stopListening()
-        dataSource = getDataSource(currentFilter).andListen()
-        return true
-    }
-
-    private fun getDataSource(currentFilter: String?): LiveData<List<HttpTransactionTuple>> {
-        val repository = RepositoryProvider.transaction()
-        return when {
-            currentFilter.isNullOrEmpty() -> repository.getSortedTransactionTuples()
-            currentFilter.isDigitsOnly() -> repository.getFilteredTransactionTuples(
-                currentFilter,
-                ""
-            )
-            else -> repository.getFilteredTransactionTuples("", currentFilter)
-        }
-    }
-
-    private fun String.isDigitsOnly(): Boolean = TextUtils.isDigitsOnly(this)
-
-    private fun LiveData<List<HttpTransactionTuple>>.stopListening() =
-        removeObservers(this@TrafficFragment)
-
-    private fun LiveData<List<HttpTransactionTuple>>.andListen(): LiveData<List<HttpTransactionTuple>> {
-        this.observe(this@TrafficFragment, Observer { tuples ->
-            tutorialView.visibility = if (tuples.isNullOrEmpty()) View.VISIBLE else View.GONE
-            trafficAdapter.submitList(tuples.map { HttpTrafficRow(it) })
-        })
-        return this
-    }
 }
